+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Changes
+* Use `kurtosis-enclave--UUID` for enclave namespace names
+
 ### Fixes
 * Fixed an issue when getting a single enclave object by its ID fails because a condition was wrong
 * Fixed a bug related to using array of reference variables for `enclaveKubernetesResources.pods` and `enclaveKubernetesResources.services` fields